+++ src/app/[locale]/(logged-out)/privacy-policy/components/PrivacyPolicy/PrivacyPolicy.tsx
